DerSPMF: moved code in src
